requirements.txt
GitPython==3.0.4
PyGithub==1.44
PyYAML==5.1.2
